# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Narvi API",
        "description": "API para la gestión de productos, usuarios, imágenes y traducciones multiidioma",
        "contact": {
            "email": "support@narvi.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/{productId}/images": {
            "post": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Subir imágenes de producto",
                "description": "Sube una o múltiples imágenes para un producto específico",
                "operationId": "4d5dd624b3772e60fd9ee0911b23ac1b",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "images[]": {
                                        "description": "Seleccione una o múltiples imágenes (jpeg, png, jpg, gif, webp - max 5MB cada una)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{productId}/images/{imageId}": {
            "delete": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Eliminar imagen de producto",
                "description": "Elimina una imagen específica de un producto",
                "operationId": "9f50ee9c8efa2eda5627847696565291",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    },
                    {
                        "name": "imageId",
                        "in": "path",
                        "description": "ID de la imagen",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagen eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Imagen eliminada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Imagen no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{productId}/images/reorder": {
            "put": {
                "tags": [
                    "Product Images"
                ],
                "summary": "Reordenar imágenes de producto",
                "description": "Actualiza el orden de visualización de las imágenes de un producto",
                "operationId": "33e6165ca4848e79625f1e6950add973",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "images"
                                ],
                                "properties": {
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID de la imagen",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order": {
                                                    "description": "Nuevo orden",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden de imágenes actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Orden de imágenes actualizado exitosamente"
                                        },
                                        "images": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese su token de autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}